refactor:핀하우스 home / 정적 데이터 ssr 컴포넌트 분리
