+++ src/features/home/index.ts
@@ -1,11 +1,11 @@
-export * from "./ui/homeContentsCard";
+export * from "./ui/client/homeContentsCard";
 export * from "./ui/homeAction/homeActionCardList";
-export * from "./ui/homeContentsCard";
-export * from "./ui/homeHeader";
-export * from "./ui/homeHero";
-export * from "./ui/homePersonalShortcutList";
-export * from "./ui/homeQuickStatsList";
-export * from "./ui/homeUrgentNoticeList";
+export * from "./ui/client/homeContentsCard";
+export * from "./ui/client/homeHeader";
+export * from "./ui/client/homeHero";
+export * from "./ui/client/homePersonalShortcutList";
+export * from "./ui/client/homeQuickStatsList";
+export * from "./ui/server/homeUrgentNoticeList.server";
 export * from "./ui/components/homeFullSheet";
 export * from "./ui/search/searchBar";
 export * from "./ui/result/homeResultSectionHeader";
