+++ src/features/home/ui/client/homeQuickStatsList.tsx
@@ -1,27 +1,16 @@
 "use client";
+
 import { CaretDown } from "@/src/assets/icons/button/caretDown";
-import { HomeFiveoclock } from "@/src/assets/icons/home/HomeFiveoclock";
-import { HomePushPin } from "@/src/assets/icons/home/homePushpin";
-import { useHomeMaxTime } from "../model/homeStore";
 import { transTime } from "@/src/shared/lib/utils";
 import { useHomeUseHooks } from "@/src/features/home/ui/homeUseHooks/homeUseHooks";
+import { useHomeMaxTime } from "@/src/features/home/model/homeStore";
 
 export const QuickStatsList = () => {
-  const { maxTime } = useHomeMaxTime();
   const { line2, line1, onSelectSection } = useHomeUseHooks();
+  const maxTime = useHomeMaxTime(s => s.maxTime);
 
   return (
-    <div className="relative grid grid-cols-2 grid-rows-[auto,1fr] rounded-2xl bg-white p-4">
-      <div className="flex items-center gap-1 text-sm text-greyscale-grey-500">
-        <HomePushPin width={17} height={17} />
-        <span>핀포인트</span>
-      </div>
-
-      <div className="flex items-center gap-1 pl-6 text-sm text-greyscale-grey-500">
-        <HomeFiveoclock width={15} height={15} />
-        <span>최대시간</span>
-      </div>
-
+    <>
       <div
         className="mt-2 flex gap-4 hover:cursor-pointer"
         onClick={() => onSelectSection("pinpoints")}
@@ -43,8 +32,6 @@ export const QuickStatsList = () => {
           </span>
         </button>
       </div>
-
-      <span className="pointer-events-none absolute bottom-3 left-1/2 top-3 w-px -translate-x-1/2 bg-greyscale-grey-200" />
-    </div>
+    </>
   );
 };
