+++ src/features/home/ui/homeUseHooks/usePersonalShortcutHooks.ts
@@ -1,3 +1,4 @@
+"use client";
 import { useOAuthStore } from "@/src/features/login/model";
 import { useEffect, useState } from "react";
 import { useRouter } from "next/navigation";
