+++ src/features/home/ui/server/homePersonalShortcutList.server.tsx
@@ -1,5 +1,5 @@
 import { PERSONAL_SHORTCUTS } from "@/src/features/home/model/model";
-import { PersonalShortcutListClient } from "@/src/features/home/ui/homePersonalShortcutList";
+import { PersonalShortcutListClient } from "@/src/features/home/ui/client/homePersonalShortcutList";
 
 export const PersonalShortcutList = () => {
   return <PersonalShortcutListClient items={PERSONAL_SHORTCUTS} />;
