+++ src/features/home/ui/server/homeUrgentNoticeList.server.tsx
@@ -1,0 +1,5 @@
+import { UrgentNoticeListClient } from "@/src/features/home/ui/client/homeUrgentNoticeList";
+
+export const HomeUrgentNoticeListServer = () => {
+  return <UrgentNoticeListClient />;
+};
