+++ src/widgets/homeSection/homeSection.tsx
@@ -1,34 +1,28 @@
-import {
-  ActionCardList,
-  HomeHeader,
-  HomeHero,
-  QuickStatsList,
-  UrgentNoticeList,
-} from "@/src/features/home";
-
+import { ActionCardList, HomeHeader } from "@/src/features/home";
 import { PageTransition } from "@/src/shared/ui/animation";
 import { PersonalShortcutList } from "@/src/features/home/ui/server/homePersonalShortcutList.server";
+import { HomeHeroServer } from "@/src/features/home/ui/server/homeHero.server";
+import { HomeQuickStatsListServer } from "@/src/features/home/ui/server/homeQuickStatsList.server";
+import { HomeUrgentNoticeListServer } from "@/src/features/home/ui/server/homeUrgentNoticeList.server";
 
 export const HomeSection = () => {
   return (
-    // <section className="relative min-h-screen w-full overflow-y-auto bg-greyscale-grey-25 text-greyscale-grey-900 scrollbar-hide"></section>
     <section className="relative min-h-screen w-full bg-greyscale-grey-25 text-greyscale-grey-900 scrollbar-hide">
       <PageTransition>
-        {/* <div className="flex flex-col pb-6"> */}
         <div className="flex flex-col">
           <div className="px-4">
             <HomeHeader />
-            <HomeHero />
+            <HomeHeroServer />
           </div>
           <div className="flex flex-col gap-3 border-b-8 border-greyscale-grey-50 px-4">
-            <QuickStatsList />
+            <HomeQuickStatsListServer />
             <ActionCardList />
           </div>
           <div className="border-b-8 border-greyscale-grey-50 p-4">
             <PersonalShortcutList />
           </div>
           <div className="px-5">
-            <UrgentNoticeList />
+            <HomeUrgentNoticeListServer />
           </div>
         </div>
       </PageTransition>
